Modified how this code is built to become a library, with one more library dependence, a lean version of which is included with this code.
Modified the Makefile to change how the tagets are built. At present, three
targets are include: the main demonstration code, the library that is the
core of this code, and a library on which the present library depends.

Streamlined the pre-processor directives that deal with C++ renaming such
that the libraries link together as expected.

A lean version of my utilities library was added to make this respository
self-sufficient.
IN 2019/05/23 "develop"