Added several getter methods to the Zone object to allow for external code to query and access (as read-only) the variables contained in the zone. This functionality is now functional for unstructured data zones.
The zone type is queried and an interger is returned, which will be enumed
in the future to trully correspond to the element type. For now a return of
zero indicates a structured dataset, positive number indicates unstructured
data, and negative one is for a zone not in the right state.

A pointer to the zone's connectivity array and the number of nodes per
element are queried with getters. The connectivity array is cast as "const"
such that the data is considered read-only.

The pointer to individual variable's data (as "double") is queried with a
getter method. The array is cast as "const" such that the data is read-only.

The driver was modified to demonstrate this functionality; the actions are
generic enough to be pretty safe.
IN 2020/08/14 "develop"